+++ src/client/App.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
+import { ErrorBoundary } from "react-error-boundary";
 import { connect } from "react-redux";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import { Router, Switch, Route } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 
 import Details from "./pages/Details";
@@ -9,46 +10,65 @@ import Profile from "./pages/Profile";
 import Search from "./pages/Search";
 import Signup from "./pages/Signup";
 
+import ErrorFallback from "./components/ErrorFallback";
 import LoadingIndicator from "./components/LoadingIndicator";
 import Navigation from "./components/Navigation";
 
+import { setError } from "./redux/actions/application";
 import { initializeApplication } from "./redux/thunks";
+
+import { history } from "./util";
 
 interface AppProps {
   handleInitializeApplication: () => void;
+  handleSetError: (error: Error, componentStack: string) => void;
 }
 
 /**
  * Application.
  */
 const App: React.SFC<AppProps> = (props: AppProps) => {
-  const { handleInitializeApplication } = props;
+  const { handleInitializeApplication, handleSetError } = props;
 
   React.useEffect(() => {
     handleInitializeApplication();
   }, []);
 
   return (
-    <Router>
+    <Router history={history}>
       <div id="app">
-        <Navigation />
-        <Switch>
-          <Route exact path="/">
-            <Search />
-          </Route>
-          <Route path="/jobs/:id">
-            <Details />
-          </Route>
-          <Route path="/login">
-            <Login />
-          </Route>
-          <Route path="/signup">
-            <Signup />
-          </Route>
-          <Route path="/profile">
-            <Profile />
-          </Route>
-        </Switch>
+        <ErrorBoundary
+          FallbackComponent={ErrorFallback}
+          onError={(error: Error, componentStack: string) => {
+            handleSetError(
+              { message: error.message, name: error.name, stack: error.stack },
+              componentStack
+            );
+          }}
+          onReset={() => {
+            history.push("/");
+            handleInitializeApplication();
+          }}
+        >
+          <Navigation />
+          <Switch>
+            <Route exact path="/">
+              <Search />
+            </Route>
+            <Route path="/jobs/:id">
+              <Details />
+            </Route>
+            <Route path="/login">
+              <Login />
+            </Route>
+            <Route path="/signup">
+              <Signup />
+            </Route>
+            <Route path="/profile">
+              <Profile />
+            </Route>
+          </Switch>
+        </ErrorBoundary>
         <LoadingIndicator />
         <ToastContainer />
       </div>
@@ -58,6 +78,8 @@ const App: React.SFC<AppProps> = (props: AppProps) => {
 
 const mapDispatchToProps = (dispatch) => ({
   handleInitializeApplication: () => dispatch(initializeApplication()),
+  handleSetError: (error: Error, componentStack: string) =>
+    dispatch(setError(error, componentStack)),
 });
 
 export default connect(null, mapDispatchToProps)(App);