+++ src/client/components/JobCard/JobCard-styled.tsx
@@ -98,19 +98,40 @@ const StyledActions = styled.div`
 `;
 
 interface SavedButtonProps {
+  disabled: boolean;
   jobIsSaved: boolean;
 }
 
 const StyledSavedButton = styled.button<SavedButtonProps>`
   background: transparent;
   border: none;
-  color: ${(props) => (props.jobIsSaved ? "#1e86ff" : "#b9bdcf")};
+  color: ${(props) =>
+    props.jobIsSaved && !props.disabled ? "#1e86ff" : "#b9bdcf"};
   margin: 0;
   padding: 0;
 
   &:hover {
-    color: #1e86ff;
-    cursor: pointer;
+    color: ${(props) => (!props.disabled ? "#1e86ff" : "#b9bdcf")};
+    cursor: ${(props) => (!props.disabled ? "pointer" : "not-allowed")};
+  }
+`;
+
+interface HideButtonProps {
+  disabled: boolean;
+  jobIsHidden: boolean;
+}
+
+const StyledHideButton = styled.button<HideButtonProps>`
+  background: transparent;
+  border: none;
+  color: ${(props) =>
+    props.jobIsHidden && !props.disabled ? "#ff1e1e" : "#b9bdcf"};
+  margin: 0;
+  padding: 0;
+
+  &:hover {
+    color: ${(props) => (!props.disabled ? "#ff1e1e" : "#b9bdcf")};
+    cursor: ${(props) => (!props.disabled ? "pointer" : "not-allowed")};
   }
 `;
 
@@ -210,4 +231,5 @@ export {
   StyledInfoContainer,
   StyledLocationContainer,
   StyledCreatedContainer,
+  StyledHideButton,
 };