+++ src/client/components/Profile/index.ts
@@ -1,6 +1,7 @@
 import ProfileDelete from "./ProfileDelete";
 import ProfileDisplay from "./ProfileDisplay";
 import ProfileEdit from "./ProfileEdit";
+import ProfileHiddenJobs from "./ProfileHiddenJobs";
 import ProfileReset from "./ProfileReset";
 import ProfileSavedJobs from "./ProfileSavedJobs";
 
@@ -8,6 +9,7 @@ export {
   ProfileDelete,
   ProfileDisplay,
   ProfileEdit,
+  ProfileHiddenJobs,
   ProfileReset,
   ProfileSavedJobs,
 };