+++ src/client/types.ts
@@ -1,6 +1,21 @@
 import { Action } from "redux";
 import { ThunkAction } from "redux-thunk";
 
+export interface AddHiddenJobErrorResponse {
+  error: string;
+}
+
+export interface AddHiddenJobSuccessResponse {
+  createdAt: string;
+  email: string;
+  hiddenJobs: [];
+  name: string;
+  savedJobs: string[];
+  updatedAt: string;
+  __v: number;
+  _id: string;
+}
+
 export interface AddSavedJobErrorResponse {
   error: string;
 }
@@ -8,6 +23,7 @@ export interface AddSavedJobErrorResponse {
 export interface AddSavedJobSuccessResponse {
   createdAt: string;
   email: string;
+  hiddenJobs: [];
   name: string;
   savedJobs: string[];
   updatedAt: string;
@@ -21,9 +37,15 @@ export interface ApplicationAction {
   payload: any;
 }
 
+export interface ApplicationError {
+  error: Error;
+  componentStack: string;
+}
+
 export interface ApplicationState {
   currentJobs: Job[];
   currentPage: number;
+  error: ApplicationError;
   fullTime: boolean;
   isLoading: boolean;
   jobDetails: Job;
@@ -61,6 +83,12 @@ export interface GetJobsErrorResponse {
 }
 
 export type GetJobsSuccessResponse = Job[];
+
+export interface GetHiddenJobsDetailsErrorResponse {
+  error: string;
+}
+
+export type GetHiddenJobsDetailsSuccessResponse = Job[];
 
 export interface GetSavedJobsDetailsErrorResponse {
   error: string;
@@ -171,6 +199,21 @@ export type NotificationType =
 
 export type PaginationNavigationType = "left" | "right";
 
+export interface RemoveHiddenJobErrorResponse {
+  error: string;
+}
+
+export interface RemoveHiddenJobSuccessResponse {
+  createdAt: string;
+  email: string;
+  hiddenJobs: string[];
+  name: string;
+  savedJobs: string[];
+  updatedAt: string;
+  __v: number;
+  _id: string;
+}
+
 export interface RemoveSavedJobErrorResponse {
   error: string;
 }
@@ -178,6 +221,7 @@ export interface RemoveSavedJobErrorResponse {
 export interface RemoveSavedJobSuccessResponse {
   createdAt: string;
   email: string;
+  hiddenJobs: string[];
   name: string;
   savedJobs: string[];
   updatedAt: string;
@@ -203,6 +247,7 @@ export interface ServerResponseError {
 export interface ServerResponseUser {
   createdAt: string;
   email: string;
+  hiddenJobs: string[];
   name: string;
   savedJobs: string[];
   updatedAt: string;
@@ -217,6 +262,7 @@ export interface SignupErrorResponse {
 export interface SignupSuccessResponse {
   createdAt: string;
   email: string;
+  hiddenJobs: string[];
   name: string;
   savedJobs: string[];
   updatedAt: string;
@@ -235,10 +281,15 @@ export interface UserState {
   editEmail: string;
   editName: string;
   email: string;
+  hiddenJobs: string[];
+  hiddenJobsCurrentPage: number;
+  hiddenJobsDetails: Job[];
+  hiddenJobsTotalPages: number;
   isDeletingProfile: boolean;
   isEditingProfile: boolean;
   isLoggedIn: false;
   isResettingPassword: boolean;
+  isViewingHiddenJobs: boolean;
   isViewingSavedJobs: boolean;
   name: string;
   password: string;